requirements.txt
torch torchvision --index-url https://download.pytorch.org/whl/cu129
pandas
numpy
pillow
matplotlib
scikit-learn
albumentations
timm
streamlit

# Code formatting tools
isort
black